Added confirmation for submissions removal / survey delete

+++ views/mcp_index.php
@@ -8,7 +8,7 @@
 			<th>Questions</th>
 			<th>Created</th>
 			<th>Updated</th>
-			<th></th>
+			<th>&nbsp;</th>
 		</tr>
 	</thead>
 	<tbody>
@@ -24,13 +24,13 @@
 						<a href="<?php echo $compile_survey_results_url . $survey['id']; ?>" class="submit">Compile Results</a>
 						<a href="<?php echo $survey_results_url . $survey['id']; ?>" class="submit">Results</a>
 						<a href="<?php echo $edit_url . $survey['id']; ?>" class="submit">Modify</a>
-						<a href="<?php echo BASE . AMP . 'C=addons_modules' . AMP . 'M=show_module_cp' . AMP . 'module=vwm_surveys' . AMP . 'method=delete_survey_submissions' . AMP . 'survey_id=' . $survey['id'];?>" class="submit">Remove All Submissions</a>
-						<a href="<?php echo $delete_url . $survey['id']; ?>" class="submit">Delete</a>
+						<a href="<?php echo BASE . AMP . 'C=addons_modules' . AMP . 'M=show_module_cp' . AMP . 'module=vwm_surveys' . AMP . 'method=delete_survey_submissions' . AMP . 'survey_id=' . $survey['id'];?>" class="submit confirm" title="Remove all survey submissions for this survey">Remove All Submissions</a>
+						<a href="<?php echo $delete_url . $survey['id']; ?>" class="submit confirm" title="Delete survey and all associated survey submissions">Delete</a>
 					</td>
 				</tr>
 			<?php endforeach; ?>
 		<?php else: ?>
-			<tr><td colspan="6">No surveys have been created yet. </td></tr>
+			<tr><td colspan="6">No surveys have been created yet.</td></tr>
 		<?php endif; ?>
 	</tbody>
 </table>
